@@ -13,7 +13,7 @@
     }
   ],
   "require": {
-    "php": ">=7.2",
+    "php": ">=7.4",
     "ext-iconv": "*",
     "ext-intl": "*",
     "ext-json": "*",
@@ -31,7 +31,8 @@
     "php-webdriver/webdriver": "^1.0.1",
     "codeception/module-filesystem": "^1.0",
     "codeception/lib-innerbrowser": "^1.0",
-    "codeception/module-rest": "*"
+    "codeception/module-rest": "*",
+    "simplesamlphp/simplesamlphp": ">=v1.17.5"
   },
   "scripts": {
   },
