@@ -36,9 +36,9 @@ class AssetTest extends \Codeception\Test\Unit
         $this->assertEquals(['_flash.php'], $result);
 
         $result = [];
-        Asset::matchPattern(dirname(__DIR__,2).'/views', '', '*/*', function($fileName) use(&$result) {
+        Asset::matchPattern(dirname(__DIR__,2).'/views', '', '*/_*', function($fileName) use(&$result) {
             $result[] = $fileName;
         });
-        $this->assertEquals(['layout/default.php', '_form/_field.php', '_form/_field_horizontal.php', '_form/_notice.php', ], $result);
+        $this->assertEquals(['_form/_field.php', '_form/_field_horizontal.php', '_form/_notice.php', ], $result);
     }
 }
