@@ -1,4 +1,5 @@
 <?php
+namespace unit;
 
 use Codeception\Test\Unit;
 use uhi67\umvc\App;
@@ -7,8 +8,7 @@ use uhi67\umvc\L10nFile;
 class L10nFileTest extends Unit
 {
     /**
-     * @var UnitTester
-     * @noinspection PhpMultipleClassDeclarationsInspection
+     * @var \UnitTester
      */
     protected $tester;
 
@@ -31,12 +31,12 @@ class L10nFileTest extends Unit
 	public function testL()
     {
 	    $this->assertInstanceOf(L10nFile::class, App::$app->l10n);
-	    $this->assertEquals(dirname(__DIR__).DIRECTORY_SEPARATOR.'_data/messages/cat1', App::$app->l10n->directory('cat1'));
-	    $this->assertEquals(dirname(__DIR__).DIRECTORY_SEPARATOR.'_data/messages/app', App::$app->l10n->directory('app'));
+	    $this->assertEquals(dirname(__DIR__).DIRECTORY_SEPARATOR.'_data/testapp/messages/cat1', App::$app->l10n->directory('cat1'));
+	    $this->assertEquals(dirname(__DIR__).DIRECTORY_SEPARATOR.'_data/testapp/messages/app', App::$app->l10n->directory('app'));
 	    $this->assertEquals(dirname(__DIR__,2).'/messages', App::$app->l10n->directory('umvc'));
 	    $this->assertEquals('hu-HU', App::$app->locale);
 	    $this->assertEquals('hu', App::$app->l10n->lang);
-	    $this->assertEquals(dirname(__DIR__).DIRECTORY_SEPARATOR.'_data/messages', App::$app->l10n->dir);
+	    $this->assertEquals(dirname(__DIR__).DIRECTORY_SEPARATOR.'_data/testapp/messages', App::$app->l10n->dir);
 	    $this->assertEquals('Fordítás', App::l('app', 'Translation', [], 'hu'));
 	    $this->assertEquals('Fordítás', App::l('app', 'Translation'));
 	    $this->assertEquals('any***', App::l('umvc/other', 'any'));
